Follow-up to r1931389: Cast 'len' to uint (it's safe because we're in a branch that alredy verified that 'len' is in boudries).
* subversion/libsvn_subr/checksum_apr.c
  (sha1_update): Add cast.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1931403 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/checksum_apr.c
@@ -95,7 +95,7 @@ sha1_update(apr_sha1_ctx_t *ctx,
       unsigned int block;
 
       if (len < UINT_MAX)
-        block = len;
+        block = (unsigned int)len;
       else
         block = UINT_MAX;
